agent-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
pyyaml==6.0.1
python-dotenv==1.0.0